pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "hass-foldingathomecontrol"
version = "2.3.1"
description = "Homeassistant Custom Component to integrate with Folding@Home"
license = "MIT"
authors = ["Kevin Stillhammer <kevin.stillhammer@gmail.com>"]
readme = "README.md"
repository = "http://github.com/eifinger/hass-foldingathomecontrol"
keywords = [
    "aiohttp",
    "asyncio",
    "foldingathome",
    "folding-at-home"
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License"
]

[tool.poetry.dependencies]
python = "^3.9"
PyFoldingAtHomeControl = "2.2.0"

[tool.poetry.dev-dependencies]
black = "22.1.0"
blacken-docs = "^1.12.1"
coverage = {version = "^6.2", extras = ["toml"]}
bandit = "^1.7.2"
codecov = "2.1.12"
codespell = "^2.1.0"
darglint = "^1.8.1"
flake8 = "4.0.1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^22.1.11"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.8.0"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.2.0"
flake8-markdown = "^0.3.0"
flake8-simplify = "^0.15.2"
mypy = "0.931"
pre-commit = "2.17.0"
pre-commit-hooks = "^4.1.0"
pytest = "6.2.5"
pytest-asyncio = "0.18.1"
pytest-cov = "^2.12.1"
pytest-homeassistant-custom-component = "0.6.11"
pytest-mock = "3.7.0"
pylint = "2.12.2"
pyupgrade = "^2.30.1"
safety = "^1.10.3"
vulture = "^2.3"
yamllint = "^1.26.3"

[tool.pytest.ini_options]
addopts = "--cov --cov-report term-missing --cov=custom_components tests"

[tool.coverage.report]
show_missing = true
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.vulture]
min_confidence = 80
paths = ["custom_components", "tests"]

[tool.semantic_release]
version_toml = "pyproject.toml:tool.poetry.version"
version_pattern = "custom_components/foldingathomecontrol/manifest.json:\"version\": \"{version}\""
version_variable = "custom_components/foldingathomecontrol/const.py:VERSION"
build_command = false
upload_to_repository = false
upload_to_release = false
commit_subject = "chore(release): {version}"
branch = "main"
